moved all routes to the routes directory
